Changed the getSecret endpoint from a GET to POST and fixed the invalid URL error.
Created the DataObject class that holds all the data for a data layer, including the file, styling, and metadata.

The Map component will now use the DataObject class to dynamically create data layers and their accessories.

getDataFiles is a new file that abstracts all of the data preparation outside of the Map component, and allows for the layers to be fetched via exported function function. Eventually, this file will be what fetches our data from GCP, packages it up in a DataObject, and sends it off to the frontend to displaying.